# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Task Tracker API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];


export const TaskStatus = {
  Pending: 'Pending',
  In_Progress: 'In Progress',
  Completed: 'Completed',
} as const;

export interface Task {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  status: TaskStatus;
  /** @nullable */
  dueDate?: string | null;
  createdAt: string;
}

export type TaskInputStatus = typeof TaskInputStatus[keyof typeof TaskInputStatus];


export const TaskInputStatus = {
  Pending: 'Pending',
  In_Progress: 'In Progress',
  Completed: 'Completed',
} as const;

export interface TaskInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  status: TaskInputStatus;
  dueDate?: string;
}

export type TaskUpdateStatus = typeof TaskUpdateStatus[keyof typeof TaskUpdateStatus];


export const TaskUpdateStatus = {
  Pending: 'Pending',
  In_Progress: 'In Progress',
  Completed: 'Completed',
} as const;

export interface TaskUpdate {
  /** @minLength 1 */
  title?: string;
  /** @nullable */
  description?: string | null;
  status?: TaskUpdateStatus;
  /** @nullable */
  dueDate?: string | null;
}

export interface TaskStats {
  total: number;
  pending: number;
  inProgress: number;
  completed: number;
}

export type ListTasksParams = {
status?: ListTasksStatus;
search?: string;
sortBy?: ListTasksSortBy;
};

export type ListTasksStatus = typeof ListTasksStatus[keyof typeof ListTasksStatus];


export const ListTasksStatus = {
  Pending: 'Pending',
  In_Progress: 'In Progress',
  Completed: 'Completed',
} as const;

export type ListTasksSortBy = typeof ListTasksSortBy[keyof typeof ListTasksSortBy];


export const ListTasksSortBy = {
  dueDate: 'dueDate',
  createdAt: 'createdAt',
  title: 'title',
} as const;