# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Task Tracker API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: tasks
    description: Task CRUD operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List all tasks
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [Pending, "In Progress", Completed]
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: sortBy
          in: query
          required: false
          schema:
            type: string
            enum: [dueDate, createdAt, title]
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Validation error

  /tasks/stats:
    get:
      operationId: getTaskStats
      tags: [tasks]
      summary: Get task statistics
      responses:
        "200":
          description: Task statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskStats"

  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a single task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
    put:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Validation error
        "404":
          description: Not found
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Task:
      type: object
      required: [id, title, status, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [Pending, "In Progress", Completed]
        dueDate:
          type: ["string", "null"]
        createdAt:
          type: string

    TaskInput:
      type: object
      required: [title, status]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        status:
          type: string
          enum: [Pending, "In Progress", Completed]
        dueDate:
          type: string

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [Pending, "In Progress", Completed]
        dueDate:
          type: ["string", "null"]

    TaskStats:
      type: object
      required: [total, pending, inProgress, completed]
      properties:
        total:
          type: integer
        pending:
          type: integer
        inProgress:
          type: integer
        completed:
          type: integer