# lib/api-zod/src/generated/types/taskInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Task Tracker API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TaskInputStatus } from './taskInputStatus';

export interface TaskInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  status: TaskInputStatus;
  dueDate?: string;
}