# lib/api-zod/src/generated/types/taskStats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Task Tracker API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TaskStats {
  total: number;
  pending: number;
  inProgress: number;
  completed: number;
}