# lib/api-zod/src/generated/types/taskUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Task Tracker API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TaskUpdateStatus } from './taskUpdateStatus';

export interface TaskUpdate {
  /** @minLength 1 */
  title?: string;
  /** @nullable */
  description?: string | null;
  status?: TaskUpdateStatus;
  /** @nullable */
  dueDate?: string | null;
}